abstracted_materials.py
